(checkers): Make pylint use a pipe
The prior implementation used source-inplace+file to make flymake-rest
generate a temporary file in the same directory as the original file and
then lint that. Sadly pylint seems to have issues with the name of the
temporary file and reports an error even when nothings wrong.

There is a --from-stdin argument pylint can take alongside the module
name which fixes this issue. However it's not optional. You omit it and
pylint will crash. For now I'm passing the buffers file-name when
available and the just _ (a valid module name for some reason) when not.