Cleanup flymake-rest-define and remove flymake-rest-parse-*
This commit deletes the previous `flymake-rest-parse-rx.el` and
`flymake-rest-parse-enumerate.el` files and adds the functionality
for them into `flymake-rest-define.el`. To use parse-rx use
`flymake-rest-define-rx` and to use parse-enumerate use
`flymake-rest-define-enumerate`.

This was done to make byte compilation cleaner. Now we don't lexically
assign any variables (such as flymake-rest-context) when it will remain
unused for a lot of syntax checkers.