define: Auto suppress errors when sending data to process
I encountered an error with jq when sending a large file to the checker.
I think when jq encounters an error it exits immediately instead of
waiting for the rest of stdin to be sent. Emacs then just keeps trying
to send data to the process which then triggers an error from the checker.
This is an issue because flymake will disable the checker when it
encounters this. For now we work around it by just catching and logging
any errors at that point in the function.